requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
nptyping==2.5.0
hydra-core==1.3.2
ultralytics==8.3.13
openvino-dev==2024.4.0
